Add global settings.yaml
