+++ ultralytics/nn/autobackend.py
@@ -10,10 +10,9 @@ import torch
 import torch.nn as nn
 from PIL import Image
 
-from ultralytics.yolo.utils import LOGGER, ROOT
+from ultralytics.yolo.utils import LOGGER, ROOT, yaml_load
 from ultralytics.yolo.utils.checks import check_requirements, check_suffix, check_version
 from ultralytics.yolo.utils.downloads import attempt_download, is_url
-from ultralytics.yolo.utils.files import yaml_load
 from ultralytics.yolo.utils.ops import xywh2xyxy
