+++ ultralytics/yolo/v8/detect/val.py
@@ -8,9 +8,8 @@ import torch
 from ultralytics.yolo.data import build_dataloader
 from ultralytics.yolo.data.dataloaders.v5loader import create_dataloader
 from ultralytics.yolo.engine.validator import BaseValidator
-from ultralytics.yolo.utils import DEFAULT_CONFIG, colorstr, ops
+from ultralytics.yolo.utils import DEFAULT_CONFIG, colorstr, ops, yaml_load
 from ultralytics.yolo.utils.checks import check_file, check_requirements
-from ultralytics.yolo.utils.files import yaml_load
 from ultralytics.yolo.utils.metrics import ConfusionMatrix, DetMetrics, box_iou
 from ultralytics.yolo.utils.plotting import output_to_target, plot_images
 from ultralytics.yolo.utils.torch_utils import de_parallel
